Record reasons why methods may not need write
Methods that are generated from attr_ do not need documentation, as well as aliased methods, and `initialize`. If any of these are present, we can flag them with `skip_write` showing that they do not need to be updated. For now we will still send them to people who are reading docs if they are documented but ones missing documentation do not need to be sent out.