[close #1] Respect .document .yardopts
`YARD.parse` was bypassing the built in `.document` and `.yardopts` support. Instead we are using `YARD::CLI::Yardoc` which has this support built in. 

We are now explicitly testing for this.

There is currently no invalidation of any kind of methods. This is done in code triage by checking for "closed" issues (issues that haven't been updated in a certain time period). We could implement something similar where we have an "archived_at" column for DocMethods and then we never send out any methods that are "archived".